requirements.txt
numpy
tensorflow==2.17.1
transformers==4.47.0
torch
python-docx
PyPDF2
langchain==0.3.23
langchain-community==0.3.21
streamlit
sentence-transformers==3.3.1
faiss-cpu
tf-keras
llama-cpp-python